Simplify statistics combination
