+++ datafusion/common/src/stats.rs
@@ -17,9 +17,9 @@
 
 //! This module provides data structures to represent statistics
 
-use std::fmt::{self, Debug, Display};
-
 use crate::{Result, ScalarValue};
+use std::fmt::{self, Debug, Display};
+use std::mem;
 
 use arrow::datatypes::{DataType, Schema, SchemaRef};
 
@@ -327,14 +327,14 @@ impl Statistics {
         }
 
         // Convert to Vec<Slot> so we can avoid copying the statistics
-        let mut columns: Vec<_> = std::mem::take(&mut self.column_statistics)
+        let mut columns: Vec<_> = mem::take(&mut self.column_statistics)
             .into_iter()
             .map(Slot::Present)
             .collect();
 
         for idx in projection {
             let next_idx = self.column_statistics.len();
-            let slot = std::mem::replace(
+            let slot = mem::replace(
                 columns.get_mut(*idx).expect("projection out of bounds"),
                 Slot::Taken(next_idx),
             );
@@ -413,6 +413,128 @@ impl Statistics {
         self.column_statistics = Statistics::unknown_column(&schema);
         self.total_byte_size = Precision::Absent;
         Ok(self)
+    }
+
+    /// Summarize many statistics into a single `Statistics` instance.
+    pub fn summarize<'a, I>(items: I, column_count: usize) -> Statistics
+    where
+        I: IntoIterator<Item = &'a Statistics>,
+    {
+        let mut col_stats_set = vec![ColumnStatistics::default(); column_count];
+        let mut num_rows = Precision::<usize>::Absent;
+        let mut total_byte_size = Precision::<usize>::Absent;
+
+        for (idx, item_stats) in items.into_iter().enumerate() {
+            if idx == 0 {
+                // First item, set values directly
+                num_rows = item_stats.num_rows;
+                total_byte_size = item_stats.total_byte_size;
+                for (index, column_stats) in
+                    item_stats.column_statistics.iter().enumerate()
+                {
+                    col_stats_set[index].null_count = column_stats.null_count;
+                    col_stats_set[index].max_value = column_stats.max_value.clone();
+                    col_stats_set[index].min_value = column_stats.min_value.clone();
+                    col_stats_set[index].sum_value = column_stats.sum_value.clone();
+                }
+                continue;
+            }
+
+            // Accumulate statistics for subsequent items
+            num_rows = add_row_stats(item_stats.num_rows, num_rows);
+            total_byte_size = add_row_stats(item_stats.total_byte_size, total_byte_size);
+
+            for (item_col_stats, col_stats) in item_stats
+                .column_statistics
+                .iter()
+                .zip(col_stats_set.iter_mut())
+            {
+                col_stats.null_count =
+                    add_row_stats(item_col_stats.null_count, col_stats.null_count);
+                set_max_if_greater(&item_col_stats.max_value, &mut col_stats.max_value);
+                set_min_if_lesser(&item_col_stats.min_value, &mut col_stats.min_value);
+                col_stats.sum_value = item_col_stats.sum_value.add(&col_stats.sum_value);
+            }
+        }
+
+        Statistics {
+            num_rows,
+            total_byte_size,
+            column_statistics: col_stats_set,
+        }
+    }
+}
+
+/// If the given value is numerically greater than the original maximum value,
+/// return the new maximum value with appropriate exactness information.
+pub fn set_max_if_greater(
+    max_nominee: &Precision<ScalarValue>,
+    max_value: &mut Precision<ScalarValue>,
+) {
+    match (&max_value, max_nominee) {
+        (Precision::Exact(val1), Precision::Exact(val2)) if val1 < val2 => {
+            *max_value = max_nominee.clone();
+        }
+        (Precision::Exact(val1), Precision::Inexact(val2))
+        | (Precision::Inexact(val1), Precision::Inexact(val2))
+        | (Precision::Inexact(val1), Precision::Exact(val2))
+            if val1 < val2 =>
+        {
+            *max_value = max_nominee.clone().to_inexact();
+        }
+        (Precision::Exact(_), Precision::Absent) => {
+            let exact_max = mem::take(max_value);
+            *max_value = exact_max.to_inexact();
+        }
+        (Precision::Absent, Precision::Exact(_)) => {
+            *max_value = max_nominee.clone().to_inexact();
+        }
+        (Precision::Absent, Precision::Inexact(_)) => {
+            *max_value = max_nominee.clone();
+        }
+        _ => {}
+    }
+}
+
+/// If the given value is numerically lesser than the original minimum value,
+/// return the new minimum value with appropriate exactness information.
+pub fn set_min_if_lesser(
+    min_nominee: &Precision<ScalarValue>,
+    min_value: &mut Precision<ScalarValue>,
+) {
+    match (&min_value, min_nominee) {
+        (Precision::Exact(val1), Precision::Exact(val2)) if val1 > val2 => {
+            *min_value = min_nominee.clone();
+        }
+        (Precision::Exact(val1), Precision::Inexact(val2))
+        | (Precision::Inexact(val1), Precision::Inexact(val2))
+        | (Precision::Inexact(val1), Precision::Exact(val2))
+            if val1 > val2 =>
+        {
+            *min_value = min_nominee.clone().to_inexact();
+        }
+        (Precision::Exact(_), Precision::Absent) => {
+            let exact_min = mem::take(min_value);
+            *min_value = exact_min.to_inexact();
+        }
+        (Precision::Absent, Precision::Exact(_)) => {
+            *min_value = min_nominee.clone().to_inexact();
+        }
+        (Precision::Absent, Precision::Inexact(_)) => {
+            *min_value = min_nominee.clone();
+        }
+        _ => {}
+    }
+}
+
+pub fn add_row_stats(
+    file_num_rows: Precision<usize>,
+    num_rows: Precision<usize>,
+) -> Precision<usize> {
+    match (file_num_rows, &num_rows) {
+        (Precision::Absent, _) => num_rows.to_inexact(),
+        (lhs, Precision::Absent) => lhs.to_inexact(),
+        (lhs, rhs) => lhs.add(rhs),
     }
 }
